+++ R/malaystopwords.R
@@ -1,0 +1,12 @@
+#' @title Data of Malay stop words
+#'
+#'
+#' @format A tibble with 512 rows and 1 variable:
+#' \describe{
+#'   \item{`stopwords`}{dbl Malay stop words}
+#' }
+
+
+
+
+"malaystopwords"