+++ tests/testthat/test_function.r
@@ -4,7 +4,7 @@ library(malaytextr)
 # For character vector
 
 test_that("stem_malay() returns a data frame when character vector is used", {
-  output_table <- stem_malay(Word = "banyaknya", dictionary = malayrootwords)
+  output_table <- stem_malay(word = "banyaknya", dictionary = malayrootwords)
   expect_s3_class(output_table, "data.frame")
 })
 
@@ -14,7 +14,7 @@ test_that("stem_malay() returns a data frame when character vector is used", {
 test_that("stem_malay() returns a data frame when a data frame is used", {
   x <- data.frame(text = c("banyaknya","sangat","terkedu", "pengetahuan"))
 
-  output_table <-  stem_malay(Word = x, dictionary = malayrootwords, col_feature1 = "text")
+  output_table <-  stem_malay(word = x, dictionary = malayrootwords, col_feature1 = "text")
   expect_s3_class(output_table, "data.frame")
 })